# backend/contracts/openapi.json
{
  "components": {
    "schemas": {
      "AgentRunKind": {
        "enum": [
          "initial",
          "rematch"
        ],
        "title": "AgentRunKind",
        "type": "string"
      },
      "AgentRunResponse": {
        "additionalProperties": false,
        "description": "Incrementally readable Agent run, safe to poll while still running.",
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "donation_id": {
            "minLength": 1,
            "title": "Donation Id",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/AgentStateEvent"
            },
            "title": "Events",
            "type": "array"
          },
          "kind": {
            "$ref": "#/components/schemas/AgentRunKind"
          },
          "result": {
            "anyOf": [
              {
                "discriminator": {
                  "mapping": {
                    "initial": "#/components/schemas/InitialAgentRunResult",
                    "rematch": "#/components/schemas/RematchAgentRunResult"
                  },
                  "propertyName": "kind"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/InitialAgentRunResult"
                  },
                  {
                    "$ref": "#/components/schemas/RematchAgentRunResult"
                  }
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "run_id": {
            "minLength": 1,
            "title": "Run Id",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/AgentRunStatus"
          },
          "transport": {
            "$ref": "#/components/schemas/AgentRunTransport"
          }
        },
        "required": [
          "run_id",
          "donation_id",
          "status",
          "kind",
          "transport",
          "events"
        ],
        "title": "AgentRunResponse",
        "type": "object"
      },
      "AgentRunStatus": {
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed"
        ],
        "title": "AgentRunStatus",
        "type": "string"
      },
      "AgentRunSummary": {
        "additionalProperties": false,
        "properties": {
          "donation_id": {
            "title": "Donation Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/AgentRunKind"
          },
          "latest_event": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentStateEvent"
              },
              {
                "type": "null"
              }
            ]
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AgentRunStatus"
          },
          "transport": {
            "$ref": "#/components/schemas/AgentRunTransport"
          }
        },
        "required": [
          "run_id",
          "donation_id",
          "status",
          "kind",
          "transport"
        ],
        "title": "AgentRunSummary",
        "type": "object"
      },
      "AgentRunTransport": {
        "enum": [
          "live",
          "replay"
        ],
        "title": "AgentRunTransport",
        "type": "string"
      },
...
      "AgentStateEvent": {
        "additionalProperties": false,
        "description": "One visible step of an Agent run.\n\nPersisted as it happens so `GET /agent-runs/{id}` can return a growing list\nwhile the run is still in flight.",
        "properties": {
          "detail": {
            "default": "",
            "title": "Detail",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          },
          "state": {
            "$ref": "#/components/schemas/AgentState"
          }
        },
        "required": [
          "sequence",
          "state",
          "label",
          "occurred_at"
        ],
        "title": "AgentStateEvent",
        "type": "object"
      },
...
      "ApiErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "field_errors": {
            "items": {
              "$ref": "#/components/schemas/ApiFieldError"
            },
            "title": "Field Errors",
            "type": "array"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          }
        },
        "required": [
          "code",
          "detail"
        ],
        "title": "ApiErrorResponse",
        "type": "object"
      },
      "ApiFieldError": {
        "additionalProperties": false,
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "title": "ApiFieldError",
        "type": "object"
      },
...
      "CapacityChangeAlert": {
        "additionalProperties": false,
        "properties": {
          "accepted_kg": {
            "minimum": 0.0,
            "title": "Accepted Kg",
            "type": "integer"
          },
          "community_id": {
            "title": "Community Id",
            "type": "string"
          },
          "community_name": {
            "title": "Community Name",
            "type": "string"
          },
          "declared_capacity_kg": {
            "minimum": 0.0,
            "title": "Declared Capacity Kg",
            "type": "integer"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "community_id",
          "community_name",
          "declared_capacity_kg",
          "accepted_kg",
          "message"
        ],
        "title": "CapacityChangeAlert",
        "type": "object"
      },
...
      "CommunityOrganisation": {
        "additionalProperties": false,
        "properties": {
          "accepted_categories": {
            "items": {
              "$ref": "#/components/schemas/FoodCategory"
            },
            "minItems": 1,
            "title": "Accepted Categories",
            "type": "array"
          },
          "community_id": {
            "title": "Community Id",
            "type": "string"
          },
          "declared_capacity_kg": {
            "description": "Whole kilograms",
            "minimum": 0.0,
            "title": "Declared Capacity Kg",
            "type": "integer"
          },
          "is_open": {
            "title": "Is Open",
            "type": "boolean"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "needs": {
            "items": {
              "$ref": "#/components/schemas/CommunityNeed"
            },
            "minItems": 1,
            "title": "Needs",
            "type": "array"
          },
          "receiving_window": {
            "$ref": "#/components/schemas/TimeWindow"
          },
          "remaining_capacity_kg": {
            "description": "Whole kilograms",
            "minimum": 0.0,
            "title": "Remaining Capacity Kg",
            "type": "integer"
          },
          "supported_storage": {
            "items": {
              "$ref": "#/components/schemas/StorageType"
            },
            "minItems": 1,
            "title": "Supported Storage",
            "type": "array"
          }
        },
        "required": [
          "community_id",
          "name",
          "location",
          "accepted_categories",
          "supported_storage",
          "needs",
          "declared_capacity_kg",
          "remaining_capacity_kg",
          "receiving_window",
          "is_open"
        ],
        "title": "CommunityOrganisation",
        "type": "object"
      },
...
      "CreateDonationRequest": {
        "additionalProperties": false,
        "description": "Donation form payload; the backend mints the authoritative donation ID.\n\n``donation_id`` remains an optional preview-only reference so the JSON\nexample in Requirement.md section 4 still validates. The create use case\nmust never treat it as the authoritative persisted identifier.\n``store_location`` is resolved from the configured store record.",
        "properties": {
          "donation_id": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Donation Id"
          },
          "handling_notes": {
            "default": "",
            "title": "Handling Notes",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/FoodItem"
            },
            "minItems": 1,
            "title": "Items",
            "type": "array"
          },
          "pickup_window": {
            "$ref": "#/components/schemas/TimeWindow"
          },
          "store_id": {
            "minLength": 1,
            "title": "Store Id",
            "type": "string"
          }
        },
        "required": [
          "store_id",
          "pickup_window",
          "items"
        ],
        "title": "CreateDonationRequest",
        "type": "object"
      },
      "CreateDonationResponse": {
        "additionalProperties": false,
        "properties": {
          "donation": {
            "$ref": "#/components/schemas/DonationRequest"
          },
          "inventory": {
            "$ref": "#/components/schemas/DonationInventory"
          }
        },
        "required": [
          "donation",
          "inventory"
        ],
        "title": "CreateDonationResponse",
        "type": "object"
      },
...
      "DeliveryOrder": {
        "additionalProperties": false,
        "properties": {
          "deadline": {
            "format": "date-time",
            "title": "Deadline",
            "type": "string"
          },
          "destination_community_id": {
            "title": "Destination Community Id",
            "type": "string"
          },
          "donation_id": {
            "title": "Donation Id",
            "type": "string"
          },
          "driver_id": {
            "title": "Driver Id",
            "type": "string"
          },
          "is_rematch": {
            "default": false,
            "title": "Is Rematch",
            "type": "boolean"
          },
          "order_id": {
            "title": "Order Id",
            "type": "string"
          },
          "origin": {
            "$ref": "#/components/schemas/Location"
          },
          "quantity_kg": {
            "description": "Whole kilograms, greater than zero",
            "exclusiveMinimum": 0.0,
            "title": "Quantity Kg",
            "type": "integer"
          },
          "route": {
            "$ref": "#/components/schemas/RouteLeg"
          },
          "status": {
            "$ref": "#/components/schemas/DeliveryStatus"
          }
        },
        "required": [
          "order_id",
          "donation_id",
          "origin",
          "destination_community_id",
          "quantity_kg",
          "driver_id",
          "route",
          "status",
          "deadline"
        ],
        "title": "DeliveryOrder",
        "type": "object"
      },
      "DeliveryStatus": {
        "enum": [
          "created",
          "driver_assigned",
          "in_transit",
          "arrived",
          "partially_accepted",
          "completed",
          "rejected"
        ],
        "title": "DeliveryStatus",
        "type": "string"
      },
...
      "DonationInventory": {
        "additionalProperties": false,
        "description": "The quantity ledger for one donation.\n\nThe four components must always sum to `total_kg`. This is asserted at every\nstate transition, not merely tested on one path -- AGENTS_FoodFlow.md 8.4\ncalls it blocker-level. The UI renders these four as a stacked bar, which is\nthe visible proof clean_code_spec 8.4 requires.",
        "properties": {
          "available_kg": {
            "description": "Whole kilograms",
            "minimum": 0.0,
            "title": "Available Kg",
            "type": "integer"
          },
          "delivered_kg": {
            "description": "Whole kilograms",
            "minimum": 0.0,
            "title": "Delivered Kg",
            "type": "integer"
          },
          "donation_id": {
            "title": "Donation Id",
            "type": "string"
          },
          "in_transit_kg": {
            "description": "Whole kilograms",
            "minimum": 0.0,
            "title": "In Transit Kg",
            "type": "integer"
          },
          "reserved_kg": {
            "description": "Whole kilograms",
            "minimum": 0.0,
            "title": "Reserved Kg",
            "type": "integer"
          },
          "total_kg": {
            "description": "Whole kilograms, greater than zero",
            "exclusiveMinimum": 0.0,
            "title": "Total Kg",
            "type": "integer"
          }
        },
        "required": [
          "donation_id",
          "total_kg",
          "available_kg",
          "reserved_kg",
          "in_transit_kg",
          "delivered_kg"
        ],
        "title": "DonationInventory",
        "type": "object"
      },
      "DonationRequest": {
        "additionalProperties": false,
        "properties": {
          "donation_id": {
            "minLength": 1,
            "title": "Donation Id",
            "type": "string"
          },
          "handling_notes": {
            "default": "",
            "title": "Handling Notes",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/FoodItem"
            },
            "minItems": 1,
            "title": "Items",
            "type": "array"
          },
          "pickup_window": {
            "$ref": "#/components/schemas/TimeWindow"
          },
          "store_id": {
            "minLength": 1,
            "title": "Store Id",
            "type": "string"
          },
          "store_location": {
            "$ref": "#/components/schemas/Location"
          }
        },
        "required": [
          "donation_id",
          "store_id",
          "store_location",
          "pickup_window",
          "items"
        ],
        "title": "DonationRequest",
        "type": "object"
      },
      "Driver": {
        "additionalProperties": false,
        "properties": {
          "driver_id": {
            "title": "Driver Id",
            "type": "string"
          },
          "is_available": {
            "title": "Is Available",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "start_location": {
            "$ref": "#/components/schemas/Location"
          },
          "vehicle_capacity_kg": {
            "description": "Whole kilograms, greater than zero",
            "exclusiveMinimum": 0.0,
            "title": "Vehicle Capacity Kg",
            "type": "integer"
          }
        },
        "required": [
          "driver_id",
          "name",
          "start_location",
          "vehicle_capacity_kg",
          "is_available"
        ],
        "title": "Driver",
        "type": "object"
      },
      "ErrorCode": {
        "description": "Expected, typed failures. Every one is a business outcome, not a crash.",
        "enum": [
          "RECIPIENT_CATEGORY_UNSUPPORTED",
          "RECIPIENT_CAPACITY_EXCEEDED",
          "STORAGE_INCOMPATIBLE",
          "RECEIVING_WINDOW_CLOSED",
          "DELIVERY_DEADLINE_MISSED",
          "DRIVER_CAPACITY_EXCEEDED",
          "DRIVER_UNAVAILABLE",
          "RECIPIENT_DECLINED_THIS_DONATION",
          "INSUFFICIENT_INVENTORY",
          "DUPLICATE_ALLOCATION",
          "QUANTITY_INTEGRITY_VIOLATION",
          "INVALID_STATE_TRANSITION",
          "AGENT_OUTPUT_INVALID",
          "AGENT_STEP_LIMIT_REACHED",
          "AGENT_TIMEOUT",
          "TOOL_VALIDATION_FAILED",
          "TOOL_NOT_FOUND",
          "TOOL_TIMEOUT",
          "TOOL_RATE_LIMITED",
          "TOOL_INVALID_RESULT",
          "TOOL_INTERNAL_FAILURE",
          "DONATION_INVALID",
          "NOT_FOUND"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
...
      "FoodItem": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "$ref": "#/components/schemas/FoodCategory"
          },
          "delivery_deadline": {
            "format": "date-time",
            "title": "Delivery Deadline",
            "type": "string"
          },
          "item_name": {
            "minLength": 1,
            "title": "Item Name",
            "type": "string"
          },
          "quantity": {
            "description": "Whole kilograms, greater than zero",
            "exclusiveMinimum": 0.0,
            "title": "Quantity",
            "type": "integer"
          },
          "storage_type": {
            "$ref": "#/components/schemas/StorageType"
          },
          "unit": {
            "const": "kg",
            "default": "kg",
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "item_name",
          "category",
          "quantity",
          "storage_type",
          "delivery_deadline"
        ],
        "title": "FoodItem",
        "type": "object"
      },
      "GlobalDashboardResponse": {
        "additionalProperties": false,
        "description": "Global operations view; quantity ledgers remain donation-scoped.",
        "properties": {
          "active_agent_decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentRunSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "active_delivery": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeliveryOrder"
              },
              {
                "type": "null"
              }
            ]
          },
          "agent_runs": {
            "items": {
              "$ref": "#/components/schemas/AgentRunSummary"
            },
            "title": "Agent Runs",
            "type": "array"
          },
          "capacity_alerts": {
            "items": {
              "$ref": "#/components/schemas/CapacityChangeAlert"
            },
            "title": "Capacity Alerts",
            "type": "array"
          },
          "capacity_change_highlight": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CapacityChangeAlert"
              },
              {
                "type": "null"
              }
            ]
          },
          "communities": {
            "items": {
              "$ref": "#/components/schemas/CommunityOrganisation"
            },
            "title": "Communities",
            "type": "array"
          },
          "deliveries": {
            "items": {
              "$ref": "#/components/schemas/DeliveryOrder"
            },
            "title": "Deliveries",
            "type": "array"
          },
          "donations": {
            "items": {
              "$ref": "#/components/schemas/DonationRequest"
            },
            "title": "Donations",
            "type": "array"
          },
          "drivers": {
            "items": {
              "$ref": "#/components/schemas/Driver"
            },
            "title": "Drivers",
            "type": "array"
          },
          "inventories": {
            "items": {
              "$ref": "#/components/schemas/DonationInventory"
            },
            "title": "Inventories",
            "type": "array"
          },
          "kpis": {
            "$ref": "#/components/schemas/DashboardKpis"
          },
          "urgent_donation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DonationRequest"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "kpis",
          "donations",
          "inventories",
          "deliveries",
          "communities",
          "drivers",
          "agent_runs",
          "capacity_alerts"
        ],
        "title": "GlobalDashboardResponse",
        "type": "object"
      },
...
      "InitialAgentRunResult": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/AllocationDecision"
          },
          "inventory": {
            "$ref": "#/components/schemas/DonationInventory"
          },
          "kind": {
            "const": "initial",
            "default": "initial",
            "title": "Kind",
            "type": "string"
          },
          "order_refs": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Order Refs",
            "type": "array"
          }
        },
        "required": [
          "decision",
          "inventory",
          "order_refs"
        ],
        "title": "InitialAgentRunResult",
        "type": "object"
      },
...
      "RematchAgentRunResult": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/RematchDecision"
          },
          "inventory": {
            "$ref": "#/components/schemas/DonationInventory"
          },
          "kind": {
            "const": "rematch",
            "default": "rematch",
            "title": "Kind",
            "type": "string"
          },
          "order_refs": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Order Refs",
            "type": "array"
          }
        },
        "required": [
          "decision",
          "inventory",
          "order_refs"
        ],
        "title": "RematchAgentRunResult",
        "type": "object"
      },
...
      "StartMatchResponse": {
        "additionalProperties": false,
        "description": "Immediate 202 response; the run continues after this is returned.",
        "properties": {
          "donation_id": {
            "minLength": 1,
            "title": "Donation Id",
            "type": "string"
          },
          "kind": {
            "const": "initial",
            "default": "initial",
            "title": "Kind",
            "type": "string"
          },
          "run_id": {
            "minLength": 1,
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "const": "queued",
            "default": "queued",
            "title": "Status",
            "type": "string"
          },
          "transport": {
            "$ref": "#/components/schemas/AgentRunTransport"
          }
        },
        "required": [
          "run_id",
          "donation_id",
          "transport"
        ],
        "title": "StartMatchResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "P2 frozen transport contract. Runtime route implementation begins in P4; all routing and ETA values are simulated.",
    "title": "FoodFlow Auckland API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/agent-runs/{run_id}": {
      "get": {
        "operationId": "getAgentRun",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Typed FoodFlow error"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Typed FoodFlow error"
          }
        },
        "summary": "Get Agent Run",
        "tags": [
          "agent-runs"
        ]
      }
    },
    "/dashboard": {
      "get": {
        "operationId": "getDashboard",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalDashboardResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Typed FoodFlow error"
          }
        },
        "summary": "Get Dashboard",
        "tags": [
          "dashboard"
        ]
      }
    },
...
    "/donations": {
      "post": {
        "operationId": "createDonation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDonationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateDonationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Typed FoodFlow error"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Typed FoodFlow error"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Typed FoodFlow error"
          }
        },
        "summary": "Create Donation",
        "tags": [
          "donations"
        ]
      }
    },
    "/donations/{donation_id}/match": {
      "post": {
        "operationId": "startDonationMatch",
        "parameters": [
          {
            "in": "path",
            "name": "donation_id",
            "required": true,
            "schema": {
              "title": "Donation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartMatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Typed FoodFlow error"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Typed FoodFlow error"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Typed FoodFlow error"
          }
        },
        "summary": "Start Match",
        "tags": [
          "agent-runs"
        ]
      }
    },
...
  }
}